update result file path and rename script
